_daga_ubs/BitcoinNotify/target/BitcoinNotify-0.0.1-SNAPSHOT/WEB-INF/classes/info/skrzypczak/dagmara/entity/MessageEncoder.java
package main.info.skrzypczak.dagmara.entity;

import javax.websocket.EncodeException;
import javax.websocket.Encoder;
import javax.websocket.EndpointConfig;

import com.google.gson.Gson;

public class MessageEncoder implements Encoder.Text<Message> {

	private static Gson gson = new Gson();

	@Override
	public String encode(Message message) throws EncodeException {
		return gson.toJson(message);
	}

	@Override
	public void init(EndpointConfig endpointConfig) {
	}

	@Override
	public void destroy() {
	}
}